added views, as well as celery tasks for views also configured wrappers
